+++ climada_petals/hazard/copernicus_interface/__init__.py
@@ -26,3 +26,4 @@ from .downloader import (
 from .heat_index import *  # This will import all functions from heat_index.py
 from .index_definitions import *  # This will import all functions from index_definitions.py
 from .seasonal_statistics import *  # This will import all functions from seasonal_statistics.py
+from .time_utils import *  # Time-related helpers (e.g. month conversion, leadtime calculation)
